+++ typedoc.config.cjs
@@ -15,6 +15,8 @@ module.exports = {
     categorizeByGroup: false,
     categoryOrder: [
         "Jobs",
+        "Reactive Values",
+        "Reactive Behaviors",
         "Resource Management",
         "Scheduling",
         "Signals",